research_assistant_agent/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.23
google-auth==2.3.3
google-auth-oauthlib==0.4.6
python-dotenv==0.19.0
openai==1.12.0 